+++ changes.md
@@ -3,6 +3,7 @@
 ## Main wip
 ## 0.0.3
 * Made SpoolCladding class.
+* Intrgrated SpoolCladding into PowerStation class
 
 ## 0.0.2
 * Upped cqterrain verison 0.1.8